test(api): Restore test runner, ref'ed in Makefile
The `./Makefile` originally referenced a `test` target, since moved to the
`./api/Makefile`, but the `./api/buildout.cfg` installs no such buildout part.  Figuring
it's better to restore that than remove it, I tried restoring that part to the API
buildout and found it easy and that most of the tests are already passing.