build(make): Reduce build times, incremental
For each step in the build process, use the real artifacts that result as targets in
`./api/Makefile`.  This prevents wasting time in the inner loop of individual
development and when the prerequisites, `... ./buildout.cfg ./src/*/setup.py`, are
maintained carefully this rarely results in surprising behavior such as something
failing because a build step wasn't detected as needed.